requirements.txt
asgiref==3.6.0
Django==4.2.1
django-model-utils==4.3.1
sqlparse==0.4.4
